pull out format table to json to own function
